restructured css styles to component folders
