+++ src/App.tsx
@@ -1,7 +1,7 @@
 // @ts-nocheck  // removes annoying parameter any type messages for file
 import { CORE_CONCEPTS } from "./data";
-import Header from './components/Header'
-import CoreConcept from './components/CoreConcept'
+import Header from './components/Header/Header'
+import CoreConcept from './components/CoreConcept/CoreConcept'
 
 function App() {
   return (
