+++ src/components/CoreConcept/CoreConcept.tsx
@@ -1,4 +1,5 @@
 // @ts-nocheck  // removes annoying parameter any type messages for file
+import './CoreConcept.css'
 
 export default function CoreConcept({image, title, description}) {
   // object converted via destructuring
