+++ src/index.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
 // import 'bootstrap/dist/css/bootstrap.css'
-import "./index.css";
-
+import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
